requirements.txt
asammdf==8.5.0
attrs==25.3.0
canmatrix==1.2
click==8.2.1
colorama==0.4.6
isal==1.7.2
lxml==5.4.0
lz4==4.4.4
numexpr==2.11.0
numpy==2.3.0
pandas==2.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
typing_extensions==4.14.0
tzdata==2025.2
